% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{pvalue}
\alias{pvalue}
\title{Test for significant difference between 2 vectors}
\usage{
pvalue(x, y)
}
\arguments{
\item{x}{First data vector}

\item{y}{Second data vector
This function tests if the distributions of the 2 vectors follow normality
and performs t-test for normal data
Wilcox test for non-normal data}
}
\value{
p-value denoting the significance
}
\description{
Test for significant difference between 2 vectors
}

