% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_pvalue.R
\name{ttest_pvalue}
\alias{ttest_pvalue}
\title{Perform t-test and get p-value}
\usage{
ttest_pvalue(x, y)
}
\arguments{
\item{x}{First data vector}

\item{y}{Second data vector}
}
\description{
Perform t-test and get p-value
}

